BicycleEfficiency.jl

A package to calculate the efficiency of a bicycle's chain drive

Frictional losses

BicycleEfficiency.P1 — Function

P1(μ1::Number, ρ::Number, T0::Number, N::Vector{Int}, ω::Number)::Float64

Compute the power loss due to friction between pins and bushings.

Arguments

μ1::Number: friction coefficient between the pin and the bushing.

ρ::Number: bushing radius.

T0::Number: free chain tension.

N::Vector{Int}: number of teeth on the front and rear sprockets (in that order).

ω::Number: pedaling cadence.

BicycleEfficiency.P2 — Function

P2(N::Vector{Int}, ω::Number, μ2::Number, T0::Number, r0::Number, γ::Number)::Float64

Compute the power loss due to chain offset

Arguments

N::Vector{Int}: number of teeth on the front and rear sprockets (in that order).

ω::Number: pedaling cadence

μ2::Number: friction coefficient between the chain and the sprocket teeth

T0::Number: free chain tension

r0::Number: radius of contact during offset operation

γ::Number: chain offset angle

BicycleEfficiency.P3 — Function

P3(μ3::Number, T0::Number, rR::Number, N::Vector{Int}, ω::Number, ψ::Number)::Float64

Compute the power loss due to interaction between rollers and sprocket teeth

Arguments

μ3::Number: friction coefficient between roller and sprocket teeth.

T0::Number: free chain tension.

rR::Number: roller radius.

N::Vector{Int}: number of teeth on the front and rear sprockets (in that order).

ω::Number: pedaling cadence

ψ::Number: absolute roller rotation angle.

BicycleEfficiency.Ptotal — Function

Ptotal(μ::Vector{Float64}, p::Number, ρ::Number, ψ::Number, rR::Number,
       T0::Number, N::Vector{Int}, ω::Number, γ::Number)::Float64

Compute the power loss due to all 3 sources (friction between pins and bushings, chain misalignment, interaction between rollers and sprocket teeth)

It is calculated as the sum of the losses from all sources, which are implemented separately in P1, P2 and P3

Arguments

μ::Vector{Float64}: vector containing the friction coefficients for cases 1, 2 and 3 (in that order).

p::Number: chain pitch.

ρ::Number: bushing radius.

rR::Number: roller radius.

T0::Number: free chain tension.

N::Vector{Int}: number of teeth on the front and rear sprockets (in that order).

ω::Number: pedaling cadence.

γ::Number: chain offset angle

BicycleEfficiency.η — Function

η(μ::Vector{Float64}, p::Number, ρ::Number, ψ::Number, rR::Number,
  T0::Number, N::Vector{Int}, ω::Number)::Float64

Compute the power transmission efficiency considering only frictional losses

Arguments

μ::Vector{Float64}: friction coefficients for cases 1, 2 and 3 (in that order).

p::Number: chain pitch.

ρ::Number: bushing radius.

ψ::Number: absolute roller rotation angle.

rR::Number: roller radius.

T0::Number: free chain tension.

N::Vector{Int}: number of teeth on the front and rear sprockets (in that order).

ω::Number: pedaling cadence.